Add code to generate affine ctEdwards addition.
Also make public the two fields u and v of EdwardsPoint, for expediency.
Otherwise it seems that, in order to generate a circuit for addition, one also
has to generate circuits to check that the points are on the curve.
Alternatively different constraint systems may be perhaps used for the different
circuits, but it seems like more work than it's worth.